requirements.txt
Django==1.6.5
Pillow==2.5.2
South==1.0
django-endless-pagination==2.0
django-extensions==1.3.10 
django-markdown-deux==1.0.4 
django-optimizations==1.0.3 
django-pure-pagination==0.2.1 
django-suit==0.2.9 
djangorestframework==2.3.14 
markdown2==2.2.1 
psycopg2==2.5.3 
raven==5.0.0 
six==1.7.3 
wsgiref==0.1.2
django-google-charts==0.1.1
